Pins objects from the latest blocks.
We want blockchain to be distributed as much as possible to increase
robustness. But in the future we could also have "thin clients" which
would not store anything permanently, or store only a subset of things
permanently, and the rest fetch on demand. But validating the chain and
computing chain luck requires all data to be available, so we will have
to change that as well for thin clients.